Масштабные измениния
- заменил шрифты, очистил от ненужных скриптов
- добавил всплывашку о 50% ускорении
- фикс добычиресурсов
- индикатор досупности апгрейда в магазине
- и рефакторинг по мелочи

+++ Assets/_Project/Scripts/UI/Panels/ShopPanel/ShopPanel.cs
@@ -97,7 +97,7 @@ public class ShopPanel : MonoBehaviour
     private void OnPurchaseStatusChanged(bool _) =>
         UpdateShopItems();
 
-    private void OnProfitLevelChanged(float _) =>
+    private void OnProfitLevelChanged(int _) =>
         UpdateShopItems();
 
     private void OnWalletChanged(float _) =>
